fix: tsc-esm library export to avoid double compilation
