+++ packages/tsc-esm/package.json
@@ -19,7 +19,10 @@
   "files": [
     "lib"
   ],
-  "main": "./lib/index.js",
+  "main": "./lib/lib.js",
+  "exports": {
+    "default": "./lib/lib.js"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/loopingz/webda.io.git"
